// front/src/app/main/views/pessoas/models/pessoa.model.ts
export interface Pessoa {
    id: number;
    nome: string;
    cpf: string;
    rg: string;
    cidade: string;
    estado: number;
    endereco: string;
    telefone: string;
    celular: string;
    email: string;
    dt_nasc: Date;
    sexo: string;
    obs: string;

    created_at: string;
    updated_at: string;
  }